requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# MLOps & Experiment Tracking
mlflow>=2.5.0
dvc>=3.0.0
optuna>=3.2.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# API & Web Services
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Monitoring & Observability
prometheus-client>=0.17.0
evidently>=0.4.0
great-expectations>=0.17.0

# Data Processing
nltk>=3.8
spacy>=3.5.0
textblob>=0.17.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API Clients for Data Ingestion
requests>=2.31.0
tweepy>=4.14.0
praw>=7.7.0
newsapi-python>=0.2.7
alpha-vantage>=2.3.1

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
locust>=2.15.0

# Data Validation
pydantic>=2.0.0
marshmallow>=3.19.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Cloud & Storage
boto3>=1.26.0
google-cloud-storage>=2.9.0
azure-storage-blob>=12.16.0
